=BG= move crud{Call,Marshall,Unmarshall} utility methods to api.go
